Better temperature and cost computing.

+++ src/main/java/com/picarious/sa/State.java
@@ -1,5 +1,5 @@
 package com.picarious.sa;
 
 public interface State {
-    double energy();
+    int energy();
 }
